+++ 3DRadSpace/Engine3DRadSpace/FFI/Graphics/Font.cpp
@@ -1,5 +1,6 @@
 #include "Font.h"
 #include "../../Graphics/Font.hpp"
+#include "../../Logging/AssetLoadingError.hpp"
 
 using namespace Engine3DRadSpace;
 using namespace Engine3DRadSpace::Graphics;
