+++ 3DRadSpace/Engine3DRadSpace/Reflection/CMakeLists.txt
@@ -12,10 +12,10 @@ add_library(3DRadSpace.Reflection SHARED
 	"FieldRepresentation.cpp"
 	"IReflectedField.hpp"
 	"IReflectedField.cpp"
-	"IReflectedField.h"
+	"../FFI/Reflection/IReflectedField.h"
 	"IReflectedFunction.hpp"
 	"IReflectedFunction.cpp"
-	"IReflectedFunction.h"
+	"../FFI/Reflection/IReflectedFunction.h"
 	"ReflectedAsset.hpp"
 	"ReflectedEnum.hpp"
 	"ReflectedEnum.cpp"
@@ -25,14 +25,14 @@ add_library(3DRadSpace.Reflection SHARED
 	"ReflectedMemberFunction.hpp"
 	"ReflectedObject.hpp"
 	"ReflectedObject.cpp"
-	"ReflectedObject.h"
+	"../FFI/Reflection/ReflectedObject.h"
 	"Reflection.hpp"
 	"UUID.hpp"
 	"UUID.cpp"
-	"UUID.h"
+	"../FFI/Reflection/UUID.h"
 	"UnknownVariable.hpp"
 	"UnknownVariable.cpp"
-	"UnknownVariable.h"
+	"../FFI/Reflection/UnknownVariable.h"
 )
 
 if (CMAKE_VERSION VERSION_GREATER 3.12)
