+++ 3DRadSpace/Engine3DRadSpace/Reflection/IReflectedFunction.cpp
@@ -1,5 +1,4 @@
 #include "IReflectedFunction.hpp"
-#include "IReflectedFunction.h"
 
 using namespace Engine3DRadSpace::Reflection;
 
@@ -56,22 +55,4 @@ const void* IReflectedFunction::DefaultValue() const
 FieldRepresentation IReflectedFunction::Representation() const
 {
 	return { {FieldRepresentationType::Function, "" }};
-}
-
-const char* E3DRSP_IReflectedFunction_Signature(E3DRSP_IReflectedFunction func)
-{
-	auto f = reinterpret_cast<IReflectedFunction*>(func);
-	return f->Signature().c_str();
-}
-
-void E3DRSP_IReflectedFunction_Invoke(
-	E3DRSP_IReflectedFunction func,
-	void* self,
-	void* outArg,
-	void** args,
-	size_t numArgs
-)
-{
-	auto f = reinterpret_cast<IReflectedFunction*>(func);
-	f->Invoke(outArg, self, std::span<void*>(args, numArgs));
 }
